Moved CallbackMixin to glue.utils.misc
